starter project up to 2.4.5
