+++ build.gradle
@@ -9,9 +9,6 @@ subprojects {
     }
     dependencies {
         compileOnly "org.projectlombok:lombok:$lombok"
-
-
-
         testCompileOnly 'org.projectlombok:lombok:1.18.12'
         testAnnotationProcessor 'org.projectlombok:lombok:1.18.12'
         testImplementation("org.junit.jupiter:junit-jupiter-api:$jupiter")
@@ -19,5 +16,11 @@ subprojects {
         compile group: 'org.apache.logging.log4j', name: 'log4j-api', version: "$log4j"
         compile group: 'org.apache.logging.log4j', name: 'log4j-core', version: "$log4j"
 
+
     }
+}
+
+dependencies {
+//    implementation 'android.arch.persistence.room:runtime:1.1.1'
+//    annotationProcessor 'android.arch.persistence.room:compiler:1.1.1'
 }
