+++ domain/build.gradle
@@ -4,6 +4,14 @@ dependencies {
 
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.10.3'
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.10.3'
+    compile 'org.springframework.boot:spring-boot-starter-web:2.0.1.RELEASE'
+    compile 'org.springframework.boot:spring-boot-starter-data-jpa:2.0.1.RELEASE'
+   //compile group: 'com.h2database', name: 'h2', version: '1.3.148'
+   runtime("com.h2database:h2:1.3.176")
+//compile group: 'javax.persistence', name: 'persistence-api', version: '1.0.2'
+   // compile group: 'org.hibernate.javax.persistence', name: 'hibernate-jpa-2.0-api', version: '1.0.0.Final'
+
+
