+++ domain/src/main/java/starterproject/domain/Doors.java
@@ -1,14 +1,16 @@
-package starterProject.domain;
+package starterproject.domain;
+
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 @Data
 @Builder
 @JsonDeserialize
 @NoArgsConstructor
 @AllArgsConstructor
-public class Wheels {
-    private int wheels;
+public class Doors {
+    private int doorNumbers;
 }
