+++ domain/src/main/java/starterproject/domain/Wheels.java
@@ -1,16 +1,14 @@
-package starterProject.domain;
-
+package starterproject.domain;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 @Data
 @Builder
 @JsonDeserialize
 @NoArgsConstructor
 @AllArgsConstructor
-public class Transmission {
-    private Boolean automatic;
+public class Wheels {
+    private int wheels;
 }
