--- domain/src/test/java/starterProject/domain/TransmissionTest.java
@@ -1,7 +1,0 @@
-package starterProject.domain;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class TransmissionTest {
-
-}
