+++ domain/src/test/java/starterproject/AppTest.java
@@ -1,10 +1,5 @@
-package starterProject;
+package starterproject;
 
-import org.junit.jupiter.api.Test;
-import starterProject.domain.Car;
-import java.util.ArrayList;
-import java.util.List;
-import static org.junit.jupiter.api.Assertions.*;
 public class AppTest {
 
 //    @Test
