+++ domain/src/test/java/starterproject/domain/TransmissionTest.java
@@ -1,0 +1,5 @@
+package starterproject.domain;
+
+class TransmissionTest {
+
+}
