+++ frontend/build.gradle
@@ -21,6 +21,16 @@ dependencies {
     testImplementation('org.springframework.boot:spring-boot-starter-test') {
         exclude group: 'org.junit.vintage', module: 'junit-vintage-engine'
         implementation 'com.google.code.gson:gson:2.8.6'
+//        compile 'org.springframework.boot:spring-boot-starter-web:2.0.1.RELEASE'
+//        compile 'org.springframework.boot:spring-boot-starter-data-jpa:2.0.1.RELEASE'
+       // compile group: 'com.h2database', name: 'h2', version: '1.3.148'
+        //runtime("com.h2database:h2:1.3.176")
+       // compile group: 'javax.persistence', name: 'persistence-api', version: '1.0.2'
+        //compile group: 'org.hibernate.javax.persistence', name: 'hibernate-jpa-2.0-api', version: '1.0.0.Final'
+
+
+
+
 
     }
 }
