--- frontend/src/main/java/olumofe/raphael/frontEnd/Service.java
@@ -1,9 +1,0 @@
-package olumofe.raphael.frontEnd;
-
-public interface Service {
-
-
-}
-
-
-
