+++ frontend/src/main/java/olumofe/raphael/frontend/Service.java
@@ -1,0 +1,15 @@
+package olumofe.raphael.frontend;
+
+
+import starterproject.domain.Car;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface Service extends JpaRepository<Car,Integer> {
+//
+//
+//
+}
+//
+//
+
