--- frontend/src/test/java/olumofe/raphael/frontEnd/CarControllerTest.java
@@ -1,55 +1,0 @@
-package olumofe.raphael.frontEnd;
-
-import net.minidev.json.JSONObject;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.ResponseEntity;
-import org.springframework.mock.web.MockServletContext;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.RequestBuilder;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.context.WebApplicationContext;
-import starterProject.domain.Car;
-import javax.servlet.ServletContext;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.net.URL;
-import static org.junit.jupiter.api.Assertions.*;
-import static org.springframework.test.web.client.match.MockRestRequestMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class CarControllerTest {
-    @LocalServerPort
-    private int port;
-    @Autowired
-    private TestRestTemplate restTemplate;
-
-   @Autowired
-    CarController carController;
-
-    @DisplayName("Test Spring @Autowired Integration")
-
-    @Test
-    public void checkObjectIsCreated() throws Exception {
-
-        ResponseEntity<String> response = restTemplate.getForEntity(
-                new URL("http://localhost:" + port + "/cars").toString(), String.class);
-       // assertEquals("Hello Controller", response.getBody());
-        assertNotNull(response.getBody());
-    }
-    }
