+++ frontend/src/test/java/olumofe/raphael/frontend/CarControllerTest.java
@@ -1,0 +1,35 @@
+package olumofe.raphael.frontend;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.ResponseEntity;
+
+import java.net.URL;
+import static org.junit.jupiter.api.Assertions.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
+//
+//@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class CarControllerTest {
+//    @LocalServerPort
+//    private int port;
+//    @Autowired
+//    private TestRestTemplate restTemplate;
+//
+//   @Autowired
+//    CarController carController;
+
+//    @DisplayName("Test Spring @Autowired Integration")
+
+//    @Test
+//    public void checkObjectIsCreated() throws Exception {
+//
+//        ResponseEntity<String> response = restTemplate.getForEntity(
+//                new URL("http://localhost:" + port + "/cars").toString(), String.class);
+//       // assertEquals("Hello Controller", response.getBody());
+//        assertNotNull(response.getBody());
+    }
+
